Added quotes around the walltime for STAR tasks
Interesting note, unquoted walltimes like 24:00:00 are evaluated as timestamps
by the YAML parser, and loaded as an integer of the number of seconds in the
timestamp. In this case 24:00:00 was 86400 seconds. Then, jetstream submits
the task to Slurm, passing -t 86400, which *almost* works. But, slurm wants
the time in minutes, unfortunately.