Update annotSeg_7102f1c.pl
AnnotSeg creates a window from which genes are selected from the ensembl gtf.  This window was found to be too large and is annotating beyond the actual region of genes.  Because it was doing a division/multiplication before a subtraction - this caused windows to be thousands of bp larger than what they actually should be.  This fix takes care of making the window size more accurate by doing the subtraction/addition before multiplication/division.